review: move git utils into its own file
