+++ src/org/integr8ly/Utils.groovy
@@ -1,8 +1,6 @@
 #!/usr/bin/groovy
 package org.integr8ly
 
-import org.jenkinsci.plugins.pipeline.modeldefinition.Utils
-
 /**
  * @param productVersion - The product version used to extract the semver from
  * @param productName - Name of the product
@@ -44,37 +42,3 @@ boolean hasNewRelease(current, latest) {
 
   return false
 }
-
-/**
- * @param tagetBranch - The git branch to merge to the base branch
- * @param baseBranch - The base git branch
- */
-void gitMerge(String targetBranch, String baseBranch) {
-  try {
-    sh "git pull --no-edit origin ${targetBranch}"
-  } catch (Exception e) {
-    error "We were unable to merge the branch '${targetBranch}' into the branch '${baseBranch}'"
-  }
-}
-
-/**
- * @param gitBranch - The git branch to push to the remote repository
- * @param forcePush - Enables force push. Defaults to false
- * @returns void
- */
-void gitPush(String gitBranch, forcePush = false) {
-  if (forcePush) {
-    // We should be using force-with-lease flag, but the agents appear to have a really old git version in them.
-    sh "git push origin ${gitBranch} --force"
-  } else {
-    sh "git push origin ${gitBranch}"
-  }
-}
-/**
- * @param untrackedFiles - Mode to specify the handling of untracked files. Options: 'no', 'normal', 'all'. Defaults to 'no'
- * @returns a string stating the changes made if the repo is dirty. Will return an empty string if no changes are found.
- */
-String gitRepoIsDirty(String untrackedFiles = 'no') {
-  return sh(returnStdout: true, script: "git status --porcelain --untracked-files=${untrackedFiles}")?.trim()
-}
-
