+++ vars/gitCommitWhenChanges.groovy
@@ -1,5 +1,5 @@
 #!/usr/bin/groovy
-import org.integr8ly.Utils
+import org.integr8ly.GitUtils
 
 /**
  * @param commitMsgTitle - Title of the commit message
@@ -7,13 +7,13 @@ import org.integr8ly.Utils
  * @returns void
  */
 def call(String commitMsgTitle, body) {
-  def utils = new Utils()
+  def gitUtils = new GitUtils()
   def changes = []
   def msgs = []
   body(msgs)
   msgs = msgs - null - ""
 
-  if(utils.gitRepoIsDirty()) {
+  if(gitUtils.gitRepoIsDirty()) {
     changes = [commitMsgTitle]
     changes << msgs.join('\n')
     changes = changes.join('\n\n')
